requirements.txt
fast-bencode==1.1.4
